ERMS-591: Refactoring OrgSector und OrgType Part6

+++ app/grails-app/services/com/k_int/kbplus/PendingChangeService.groovy
@@ -151,44 +151,54 @@ def performAccept(change,httpRequest) {
         }      
         def updateProp = target_object.customProperties.find{it.type.name == changeDoc.name}
         if(updateProp){
-          switch (changeDoc.event){
-            case "CustomProperty.deleted":
-              log.debug("Deleting property ${updateProp.type.name} from ${parsed_change_info.changeTarget}")
-              target_object.customProperties.remove(updateProp)
-              updateProp.delete()
-              break;
-            case "CustomProperty.updated":
-              log.debug("Update custom property ${updateProp.type.name}")
-              if(changeDoc.type == RefdataValue.toString()){
-                  def propertyDefinition = PropertyDefinition.findByName(changeDoc.name)
-                  def newProp =  RefdataCategory.lookupOrCreate(propertyDefinition.refdataCategory,changeDoc.new)
-                  updateProp."${changeDoc.prop}" = newProp                
-              }else{
-                updateProp."${changeDoc.prop}" = 
-                  updateProp.parseValue("${changeDoc.new}", changeDoc.type)
-              }
-              log.debug("Setting value for ${changeDoc.name}.${changeDoc.prop} to ${changeDoc.new}")
-              updateProp.save()          
-              break;
-            default:
-              log.error("ChangeDoc event '${changeDoc.event}'' not recognized.")          
-          }
+
+            if (changeDoc.event.endsWith('CustomProperty.deleted')) {
+
+                log.debug("Deleting property ${updateProp.type.name} from ${parsed_change_info.changeTarget}")
+                target_object.customProperties.remove(updateProp)
+                updateProp.delete()
+            }
+            else if (changeDoc.event.endsWith('CustomProperty.updated')) {
+
+                log.debug("Update custom property ${updateProp.type.name}")
+                if (changeDoc.type == RefdataValue.toString()){
+                    def propertyDefinition = PropertyDefinition.findByName(changeDoc.name)
+                    def newProp =  RefdataCategory.lookupOrCreate(propertyDefinition.refdataCategory,changeDoc.new)
+                    updateProp."${changeDoc.prop}" = newProp
+                }
+                else {
+                    updateProp."${changeDoc.prop}" =
+                            updateProp.parseValue("${changeDoc.new}", changeDoc.type)
+                }
+                log.debug("Setting value for ${changeDoc.name}.${changeDoc.prop} to ${changeDoc.new}")
+                updateProp.save()
+            }
+            else {
+                log.error("ChangeDoc event '${changeDoc.event}'' not recognized.")
+            }
+
         }else{
           if(changeDoc.propertyOID){
-            def propertyType = genericOIDService.resolveOID(changeDoc.propertyOID).type
-            def newProperty = PropertyDefinition.createGenericProperty(PropertyDefinition.CUSTOM_PROPERTY, target_object, propertyType)
+            def propertyType = genericOIDService.resolveOID(changeDoc.propertyOID)?.type
+              if (propertyType) {
 
-            if(changeDoc.type == RefdataValue.toString()){
-              def originalRefdata = genericOIDService.resolveOID(changeDoc.propertyOID).refValue;
-              log.debug("RefdataCategory ${propertyType.refdataCategory}")
-              def copyRefdata = RefdataCategory.lookupOrCreate(propertyType.refdataCategory,changeDoc.new)
-              newProperty."${changeDoc.prop}" = copyRefdata
-            }else{
-              newProperty."${changeDoc.prop}" = AbstractProperty.parseValue("${changeDoc.new}", changeDoc.type)
-            }
-            newProperty.save()
-            log.debug("New CustomProperty ${newProperty.type.name} created for ${target_object}")
-            
+
+                  def newProperty = PropertyDefinition.createGenericProperty(PropertyDefinition.CUSTOM_PROPERTY, target_object, propertyType)
+
+                  if (changeDoc.type == RefdataValue.toString()) {
+                      def originalRefdata = genericOIDService.resolveOID(changeDoc.propertyOID).refValue;
+                      log.debug("RefdataCategory ${propertyType.refdataCategory}")
+                      def copyRefdata = RefdataCategory.lookupOrCreate(propertyType.refdataCategory, changeDoc.new)
+                      newProperty."${changeDoc.prop}" = copyRefdata
+                  } else {
+                      newProperty."${changeDoc.prop}" = AbstractProperty.parseValue("${changeDoc.new}", changeDoc.type)
+                  }
+                  newProperty.save()
+                  log.debug("New CustomProperty ${newProperty.type.name} created for ${target_object}")
+              }
+              else {
+                  log.error("Custom property change, but can not relove propertyType: ${parsed_change_info}")
+              }
           }else{
             log.error("Custom property change, but changedoc is missing propertyOID: ${parsed_change_info}")
           }
